% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\docType{methods}
\name{as.data.frame,CollapsedVCF-method}
\alias{as.data.frame,CollapsedVCF-method}
\title{Convert CollapsedVCF to data.frame}
\usage{
\S4method{as.data.frame}{CollapsedVCF}(x, row.names = NULL,
  optional = FALSE, ...)
}
\description{
This method converts an \linkS4class{CollapsedVCF} to a data.frame.
}
\examples{
require(VariantAnnotation)
vcffile = system.file(package='VariantAnnotation',path='extdata/chr22.vcf.gz')
v = readVcf(vcffile,'hg19')
head(as.data.frame(v))

}

