% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\docType{methods}
\name{as.data.frame,ExpandedVCF-method}
\alias{as.data.frame,ExpandedVCF-method}
\title{Convert ExpandedVCF to data.frame}
\usage{
\S4method{as.data.frame}{ExpandedVCF}(x, row.names = NULL, optional = FALSE,
  ...)
}
\description{
This method converts an \linkS4class{ExpandedVCF} to a data.frame.
}
\examples{
require(VariantAnnotation)
vcffile = system.file(package='VariantAnnotation',path='extdata/chr22.vcf.gz')
v = readVcf(vcffile,'hg19')
ve = expand(v)
class(ve)
head(as.data.frame(v))

}
