% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\docType{methods}
\name{as.json,VCF-method}
\alias{as.json,VCF-method}
\title{Convert VCF object to json}
\usage{
\S4method{as.json}{VCF}(x, ...)
}
\description{
This method converts an \linkS4class{VCF} to a JSON string.
}
\examples{
require(VariantAnnotation)
vcffile = system.file(package='VariantAnnotation',path='extdata/chr22.vcf.gz')
v = readVcf(vcffile,'hg19')
as.json(head(v,3),pretty=TRUE)

}

