% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SV.R
\name{structuralVariantVCFToGRanges}
\alias{structuralVariantVCFToGRanges}
\title{Convert structural variant VCF to GRanges}
\usage{
structuralVariantVCFToGRanges(vcf, deletion_size = Inf)
}
\arguments{
\item{vcf}{\code{\link[VariantAnnotation]{VCF}} object encoding
variants in VCF format (see
\link{\url{http://samtools.github.io/hts-specs/VCFv4.2.pdf}}}

\item{deletion_upper_limit}{integer(1) giving the upper limit on
size of deletions to include. The idea is to treat focal
deletions a little differently than large-scale deletions.}
}
\value{
\code{\link[GenomicRanges]{GRanges}} object with the
    structural variants coded as in Details
}
\description{
This function is just an opinionated approach to convert a VCF file
that encodes SVs into a GRanges object. The rationale for this
function is to be identify locations in the genome that disrupt
genes (or other regions of interest). An overlap of the returned
regions from this function with genes will result in a list of
putatively disrupted genes.
}
\details{
BND (break ends) are converted to single base pair
\code{GRanges}. DUPs are considered as having two breaks and are
converted to two individual single base pair regions. DELs are
considered as contiguous regions and are converted to one GRanges
object with \code{start} and \code{end} as specified in the VCF
file.
}
